Data access layer configuration. An interface PhoneRepository is created extending JpaRepository with a type Phone, and it's id type specified.
In PhoneService, an instance of PhoneRepository phoneRepository is made, and it's constructor is made. The constructor is annotated with @Autowired as we did for the constructor in PhoneService.

At our method, List<Phone> getPhones(), we return items returned by the function findAll() instead of the static list.
This method grabs from a database we specified.